Array implementation with Modularity
