+++ checker.cpp
@@ -1,90 +1,23 @@
 #include <assert.h>
 #include <iostream>
+#include "initiateRangeChecker.h"
+#include "checkVitals.h"
 
-class AlertInterface
+void AlertSMS(const char *name, const char *msg)
 {
-  public:
-    virtual void raiseAlert(const char* vitalName, const char* level) = 0;
+  std::cout << "SMS: " << name << " " << msg << std::endl;
 };
 
-class AlertWithSMS: public AlertInterface
+int main()
 {
-  public:
-    void raiseAlert(const char* vitalName, const char* level) override
-    {
-      std::cout << "SMS: " << vitalName << " " << level << std::endl;
-    }
-};
+  //Vital-Integrator
+  VitalCheckerInstance *vitalCheckers = initiateRangeChecker();
+  
+  //alert messages
+  AlertMessages* alerts = checkVitals(vitalCheckers);
 
-class AlertWithSound: public AlertInterface
-{
-  public:
-    void raiseAlert(const char* vitalName, const char* level) override
-    {
-      std::cout << "Sound: " << vitalName << " " << level << std::endl;  //'what'
-    }
-};
-
-class AlertIntegrator : public AlertInterface
-{
-  private:
-    AlertWithSMS smsAlerter;
-    AlertWithSound soundAlerter;
-  public:
-    void raiseAlert(const char* vitalName, const char* level) override
-    {
-      smsAlerter.raiseAlert(vitalName, level);
-      soundAlerter.raiseAlert(vitalName, level);
-    }
-};
-
-class RangeChecker
-{
-  private:
-    int lower;
-    int upper;
-    const char* vitalName;
-    AlertInterface* alerter;
-  public:
-    RangeChecker(const char* name, int low, int up, AlertInterface* alerterPtr)
-    {
-      vitalName = name;
-      lower = low;
-      upper = up;
-      alerter = alerterPtr;
-    }
-    void checkAgainstRange(float value)
-    {
-      if(value < lower) {
-        alerter->raiseAlert(vitalName, "too low");  //'when'
-      } else if(value > upper) {
-        alerter->raiseAlert(vitalName, "too high");
-      }
-    }
-};
-
-class VitalsIntegrator
-{
-  private:
-    RangeChecker bpmChecker, spo2Checker, respChecker;
-  public:
-    VitalsIntegrator(AlertInterface* alertPtr): 
-      bpmChecker("pulse rate", 70, 150, alertPtr),
-      spo2Checker("spo2", 90, 101, alertPtr),
-      respChecker("resp rate", 30, 95, alertPtr)
-    {}
-    void checkAllVitals(float bpm, float spo2, float respRate)
-    {
-      bpmChecker.checkAgainstRange(bpm);
-      spo2Checker.checkAgainstRange(spo2);
-      respChecker.checkAgainstRange(respRate);
-    }
-};
-
-int main() {
-  AlertIntegrator alerter;
-  VitalsIntegrator vitals(&alerter);
-
-  vitals.checkAllVitals(80, 95, 60);
-  vitals.checkAllVitals(60, 90, 40);
+  for (int i = 0; i < numberOfInputsInConfig; i++)
+  {
+    AlertSMS(alerts[i].vitalName, alerts[i].message);
+  }
 }
