+++ initiateRangeChecker.cpp
@@ -1,10 +1,14 @@
 #include "initiateRangeChecker.h"
 
-void initiateRangeChecker(const InputVitalInfo* inputConfig){
-    for (int i = 0; i < numberOfInputsInConfig; i++)
+VitalCheckerInstance vc[numberOfInputsInConfig];
+
+VitalCheckerInstance *initiateRangeChecker()
+{
+  for (int i = 0; i < numberOfInputsInConfig; i++)
   {
     RangeChecker *range = new RangeChecker(inputConfig[i].lower, inputConfig[i].upper);
-    vitalCheckers[i].vitalName = inputConfig[i].vitalName;
-    vitalCheckers[i].range = range;
+    vc[i].vitalName = inputConfig[i].vitalName;
+    vc[i].range = range;
   }
+  return vc;
 }
